Refactor: Move LCNC reduction engine to root reduction package
- Moved all files from `lcnc/reduction/` to `reduction/` in both `commonMain` and `commonTest`.
- Updated package declarations and imports across the codebase.
- Replaced `Capability.Mesh` with `Capability.Sctp` and `Capability.ModelMux` with `Capability.Model` in `ReducerRegistry` and related tests to match current `Capability` sealed class definitions.
