+++ src/commonMain/kotlin/borg/trikeshed/reduction/ForgeReducers.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 /**
  * Forge-specific Folder/Merger implementations extracted from ForgeWorkspaceImpl.
