+++ src/commonMain/kotlin/borg/trikeshed/reduction/LcncReduction.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 import borg.trikeshed.lib.*
