+++ src/commonMain/kotlin/borg/trikeshed/reduction/ReducerRegistry.kt
@@ -1,8 +1,8 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 import borg.trikeshed.context.nuid.Capability
 
-val Capability.category: String
+val borg.trikeshed.context.nuid.Capability.category: String
     get() = when (this) {
         is Capability.Process -> "process"
         is Capability.Cas -> "cas"
