+++ src/commonTest/kotlin/borg/trikeshed/reduction/ReducerRegistryTest.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.lcnc.reduction
+package borg.trikeshed.reduction
 
 import borg.trikeshed.context.nuid.Capability
 import kotlin.test.Test
